requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai>=1.40.0
pypdf2
python-docx
openpyxl
python-pptx
olefile
requests
qdrant-client>=1.7.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
tiktoken
boto3
botocore
httpx
